Refactor settings resource

+++ backend/app/BusinessLogic/Services/AcceptedDomainService.php
@@ -8,7 +8,6 @@ use Illuminate\Support\Facades\Auth;
 use App\Library\ApiResponse;
 use App\Models\AcceptedDomain;
 use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Support\Facades\Route;
 
 /**
  * AcceptedDomainService is a service class the will implement all the methods from the AcceptedDomainInterface contract and will handle the business logic.
@@ -20,9 +19,6 @@ class AcceptedDomainService implements AcceptedDomainInterface
     protected $modelName;
     protected $apiResponse;
     protected $currentRouteName;
-    protected $availableRoutes = [
-        'index'
-    ];
 
     /**
      * Instantiate the variables that will be used to get the model and table name as well as the table's columns.
@@ -30,9 +26,6 @@ class AcceptedDomainService implements AcceptedDomainInterface
      */
     public function __construct()
     {
-        $this->currentRouteName = Route::current()->getName();
-        dd($this->currentRouteName);
-
         $this->modelName = new AcceptedDomain();
         $this->apiResponse = new ApiResponse();
     }
@@ -117,13 +110,16 @@ class AcceptedDomainService implements AcceptedDomainInterface
      */
     public function handleDestroy($id)
     {
-        $apiDisplaySingleRecord = $this->modelName->fetchSingleRecord($id);
-        if ($apiDisplaySingleRecord && $apiDisplaySingleRecord->isNotEmpty())
-        {
-            $this->modelName->deleteRecord($id);
+        if (Auth::user() && Auth::user()->role_id === 1) {
+            $apiDisplaySingleRecord = $this->modelName->fetchSingleRecord($id);
+            if ($apiDisplaySingleRecord && $apiDisplaySingleRecord->isNotEmpty()) {
+                $this->modelName->deleteRecord($id);
+            }
+            $apiDeleteRecord = $this->apiResponse->generateApiResponse($apiDisplaySingleRecord, 'delete');
+            return $apiDeleteRecord;
+        } else {
+            return $this->apiResponse->generateApiResponse(null, 'not_allowed');
         }
-        $apiDeleteRecord = $this->apiResponse->generateApiResponse($apiDisplaySingleRecord, 'delete');
-        return $apiDeleteRecord;
     }
 
     /**
